Add `either` function
From the exercise to write a function that accepts the atom `success` or
a tuple (pair in this case) of `{error, Message}'. If it recieves the
atom `success`, it should output the string `"success"`.  If it recieves
a pair whose first element is the atom `error`, it should output the
string `"error: <Message>"`.